requirements.txt
tensorflow
numpy
scikit-learn
matplotlib
Pillow
tqdm